requirements.txt
# local package
-e .

# external requirements
Sphinx
coverage
awscli
flake8
python-dotenv
tika
PdfReader
textacy
absl-py
aiobotocore
aiohttp
aioitertools
aiosignal
asttokens
astunparse
async-timeout
attrs
autopep8
backcall
blis
botocore
cachetools
catalogue
certifi
cffi
charset-normalizer
chocolate
click
colorama
confection
cryptography
csvwlib
cymem
debugpy
decorator
duckdb
elementpath
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints
executing
falcon
filelock
flatbuffers
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.4.0/fr_core_news_md-3.4.0-py3-none-any.whl
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.4.0/fr_core_news_sm-3.4.0-py3-none-any.whl
fr-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/fr_dep_news_trf-3.4.0/fr_dep_news_trf-3.4.0-py3-none-any.whl
frozenlist
fsspec
gast
gcsfs
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
greenlet
grpcio
h5py
html5lib
huggingface-hub
icecream
idna
importlib-metadata
ipykernel
ipython
isodate
jedi
Jinja2
jmespath
joblib
jsonpath-python
jsonpickle
jupyter-core
jupyter_client
keras
Keras-Preprocessing
kglab
langcodes
language-tags
libclang
Markdown
MarkupSafe
matplotlib-inline
morph-kgc
multidict
murmurhash
nest-asyncio
networkx
nltk
numpy
oauthlib
opt-einsum
Owlready2
owlrl
oxrdflib
packaging
pandas
parso
pathy
patsy
pickleshare
preshed
prettytable
prompt-toolkit
protobuf
pslpython
psutil
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycodestyle
pycparser
pydantic
Pygments
pynvml
pyoxigraph
pyparsing
pyshacl
python-dateutil
pytz
pyvis
pywin32
PyYAML
pyzmq
rdflib
rdflib-jsonld
regex
requests
requests-oauthlib
rsa
s3fs
scikit-learn
scipy
sentencepiece
six
smart-open
spacy
spacy-alignments
spacy-legacy
spacy-loggers
spacy-transformers
sql-metadata
SQLAlchemy
sqlparse
srsly
stack-data
statsmodels
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-hub
tensorflow-intel
tensorflow-io-gcs-filesystem
tensorflow-text
termcolor
thinc
threadpoolctl
tokenizers
toml
torch
tornado
tqdm
traitlets
transformers
typer
typing_extensions
uritemplate
urllib3
urlpath
wasabi
wcwidth
webencodings
Werkzeug
wrapt
yarl
zipp